requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
pillow

tensorflow
torch

xgboost

pyarrow

streamlit